D3.js Scatter Plot Chart
A scatter plot chart that visualizes consumption of a single device for a 24hrs. period of time. It reads the data from a JSON object and ideally should be on a separate file. 
The code is based on Mike Bostock's Scatterplot visualisation which can be found here: crowdgrader.org/crowdgrader/default/index